--- a/.kitchen.yml
+++ b/.kitchen.yml
@@ -3,66 +3,66 @@ driver:
   use_sudo: false
 
 platforms:
-  - name: centos-7
-    driver_config:
-      run_command: /usr/sbin/init
-      privileged: true
-      platform: centos
-      provision_command:
-        - yum install -y epel-release
-        - yum-config-manager --add-repo https://releases.ansible.com/ansible/rpm/release/epel-7-x86_64 ansible
-        - rpm --import https://releases.ansible.com/keys/RPM-GPG-KEY-ansible-release.pub
-        - yum install -y ansible-2.6.0 initscripts rsyslog
-        - sed -i 's/#UseDNS yes/UseDNS no/g' /etc/ssh/sshd_config
-        - sed -i 's/UsePAM yes/UsePAM no/g' /etc/ssh/sshd_config
-        - sed -i 's/GSSAPIAuthentication yes/GSSAPIAuthentication no/g' /etc/ssh/sshd_config
-        - sed -i 's/#LogLevel INFO/LogLevel DEBUG/g' /etc/ssh/sshd_config
-        - systemctl enable rsyslog.service
-        - systemctl enable sshd.service
+- name: centos-7
+  driver_config:
+    run_command: /usr/sbin/init
+    privileged: true
+    platform: centos
+    provision_command:
+    - yum install -y epel-release
+    - yum-config-manager --add-repo https://releases.ansible.com/ansible/rpm/release/epel-7-x86_64 ansible
+    - rpm --import https://releases.ansible.com/keys/RPM-GPG-KEY-ansible-release.pub
+    - yum install -y ansible-2.6.0 initscripts rsyslog
+    - sed -i 's/#UseDNS yes/UseDNS no/g' /etc/ssh/sshd_config
+    - sed -i 's/UsePAM yes/UsePAM no/g' /etc/ssh/sshd_config
+    - sed -i 's/GSSAPIAuthentication yes/GSSAPIAuthentication no/g' /etc/ssh/sshd_config
+    - sed -i 's/#LogLevel INFO/LogLevel DEBUG/g' /etc/ssh/sshd_config
+    - systemctl enable rsyslog.service
+    - systemctl enable sshd.service
 
-  - name: debian-9
-    driver_config:
-      platform: debian
-      provision_command:
-        - apt-get install -y apt-utils libffi-dev python-pycparser python-apt python-dev python-pip
-        - pip install ansible==2.6.0.0
+- name: debian-9
+  driver_config:
+    platform: debian
+    provision_command:
+    - apt-get install -y apt-utils libffi-dev python-pycparser python-apt python-dev python-pip
+    - pip install ansible==2.6.0.0
 
-  - name: fedora-27
-    driver_config:
-      image: fedora:27
-      run_command: /usr/sbin/init
-      privileged: true
-      provision_command:
-        - mkdir -p /var/run/cassandra
-        - dnf install -y rsyslog iproute yum-utils initscripts
-        - yum-config-manager --add-repo https://releases.ansible.com/ansible/rpm/release/epel-7-x86_64 ansible
-        - rpm --import https://releases.ansible.com/keys/RPM-GPG-KEY-ansible-release.pub
-        - dnf install -y ansible-2.6.0
-        - sed -i 's/#UseDNS yes/UseDNS no/g' /etc/ssh/sshd_config
-        - sed -i 's/UsePAM yes/UsePAM no/g' /etc/ssh/sshd_config
-        - sed -i 's/GSSAPIAuthentication yes/GSSAPIAuthentication no/g' /etc/ssh/sshd_config
-        - sed -i 's/#LogLevel INFO/LogLevel DEBUG/g' /etc/ssh/sshd_config
-        - systemctl enable rsyslog.service
-        - systemctl enable sshd.service
+- name: fedora-27
+  driver_config:
+    image: fedora:27
+    run_command: /usr/sbin/init
+    privileged: true
+    provision_command:
+    - mkdir -p /var/run/cassandra
+    - dnf install -y rsyslog iproute yum-utils initscripts
+    - yum-config-manager --add-repo https://releases.ansible.com/ansible/rpm/release/epel-7-x86_64 ansible
+    - rpm --import https://releases.ansible.com/keys/RPM-GPG-KEY-ansible-release.pub
+    - dnf install -y ansible-2.6.0
+    - sed -i 's/#UseDNS yes/UseDNS no/g' /etc/ssh/sshd_config
+    - sed -i 's/UsePAM yes/UsePAM no/g' /etc/ssh/sshd_config
+    - sed -i 's/GSSAPIAuthentication yes/GSSAPIAuthentication no/g' /etc/ssh/sshd_config
+    - sed -i 's/#LogLevel INFO/LogLevel DEBUG/g' /etc/ssh/sshd_config
+    - systemctl enable rsyslog.service
+    - systemctl enable sshd.service
 
-  - name: ubuntu-16.04
-    driver_config:
-      platform: ubuntu
-      provision_command:
-        - apt-get install -y apt-utils libffi-dev python-pycparser python-apt python-dev python-pip
-        - pip install ansible==2.6.0.0
+- name: ubuntu-16.04
+  driver_config:
+    platform: ubuntu
+    provision_command:
+    - apt-get install -y apt-utils libffi-dev python-pycparser python-apt python-dev python-pip
+    - pip install ansible==2.6.0.0
 
-  - name: ubuntu-18.04
-    driver_config:
-      platform: ubuntu
-      provision_command:
-        - mkdir -p /var/run/sshd
-        - apt-get install -y apt-utils iproute2 libffi-dev python-pycparser python-apt python-dev python-pip
-        - pip install ansible==2.6.0.0
+- name: ubuntu-18.04
+  driver_config:
+    platform: ubuntu
+    provision_command:
+    - mkdir -p /var/run/sshd
+    - apt-get install -y apt-utils iproute2 libffi-dev python-pycparser python-apt python-dev python-pip
+    - pip install ansible==2.6.0.0
 
 provisioner:
   additional_copy_path:
-    - test
+  - test
   ansible_inventory: test/inventory
   ansible_verbose: false
   ansible_verbosity: 1
@@ -78,12 +78,12 @@ verifier:
   name: inspec
 
 suites:
-  - name: default
-    provisioner:
-      idempotency_test: true
-      playbook: test/test.yml
+- name: default
+  provisioner:
+    idempotency_test: true
+    playbook: test/test.yml
 
-  - name: check
-    provisioner:
-      ansible_check: true
-      playbook: test/test.yml
+- name: check
+  provisioner:
+    ansible_check: true
+    playbook: test/test.yml