--- a/.travis.yml
+++ b/.travis.yml
@@ -28,10 +28,8 @@ install:
 
 script:
   - make lint
-  - bundle exec kitchen list
   - bundle exec kitchen create $INSTANCE || true # A minor schlonk to avoid failure during kitchen build.
   - bundle exec kitchen test $INSTANCE
-  - make install_current_version
 
 notifications:
   webhooks: https://galaxy.ansible.com/api/v1/notifications/