--- a/Makefile
+++ b/Makefile
@@ -13,7 +13,3 @@ lint:
 	  tasks/directory.yml \
 	  tasks/main.yml \
 	  test/test.yml
-
-install_current_version:
-	rm -rf /tmp/roles
-	ansible-galaxy install --roles-path /tmp/roles locp.cassandra