--- a/test/ansible-review.cfg
+++ b/test/ansible-review.cfg
@@ -1,2 +1,2 @@
 [rules]
-standards = tests/ansible-review
+standards = test/ansible-review