--- a/test/integration/default/inspec/cluster_name_spec.rb
+++ b/test/integration/default/inspec/cluster_name_spec.rb
@@ -1,9 +1,7 @@
-describe 'Common Check Cluster Name' do
-  @script = 'SELECT cluster_name FROM local;'
-  @creds = '-u cassandra -p cassandra'
-  @cqlsh = 'cqlsh $( hostname -i ) -k system'
+@script = 'SELECT cluster_name FROM local;'
+@creds = '-u cassandra -p cassandra'
+@cqlsh = 'cqlsh $( hostname -i ) -k system'
 
-  describe command("#{@cqlsh} #{@creds} -e '#{@script}'") do
-    its(:stdout) { should match(/MyCassandraCluster/) }
-  end
+describe command("#{@cqlsh} #{@creds} -e '#{@script}'") do
+  its(:stdout) { should match(/MyCassandraCluster/) }
 end