--- a/test/integration/default/inspec/cluster_name_spec.rb
+++ b/test/integration/default/inspec/cluster_name_spec.rb
@@ -1,6 +1,7 @@
 @script = 'SELECT cluster_name FROM local;'
 @creds = '-u cassandra -p cassandra'
-@cqlsh = 'cqlsh $( hostname -i ) -k system'
+@ip = command('hostname -I').stdout.strip
+@cqlsh = "cqlsh #{@ip} -k system"
 
 describe command("#{@cqlsh} #{@creds} -e '#{@script}'") do
   its(:stdout) { should match(/MyCassandraCluster/) }