--- a/test/integration/default/inspec/cluster_name_spec.rb
+++ b/test/integration/default/inspec/cluster_name_spec.rb
@@ -5,4 +5,5 @@
 
 describe command("#{@cqlsh} #{@creds} -e '#{@script}'") do
   its(:stdout) { should match(/MyCassandraCluster/) }
+  its(:stderr) { should match(//) }
 end