--- a/test/integration/default/inspec/config_file_spec.rb
+++ b/test/integration/default/inspec/config_file_spec.rb
@@ -1,0 +1,10 @@
+config_file = if os.family == 'redhat'
+                '/etc/cassandra/conf/cassandra.yaml'
+              else
+                '/etc/cassandra/cassandra.yaml'
+              end
+
+describe file(config_file) do
+  it { should be_file }
+  it { should contain 'cluster_name: MyCassandraCluster' }
+end