--- a/test/integration/default/inspec/package_spec.rb
+++ b/test/integration/default/inspec/package_spec.rb
@@ -1,5 +1,3 @@
-require_relative '../spec_helper'
-
 describe 'Common Package' do
   describe package('cassandra') do
     it { should be_installed }