--- a/test/integration/default/inspec/rackdc_spec.rb
+++ b/test/integration/default/inspec/rackdc_spec.rb
@@ -6,6 +6,6 @@ rackdc_file = if os.family == 'redhat'
 
 describe file(rackdc_file) do
   it { should be_file }
-  it { should contain 'dc=DC1' }
-  it { should contain 'rack=RACK1' }
+  its('content') { should match(/dc=DC1/) }
+  its('content') { should match(/rack=RACK1/) }
 end