--- a/test/integration/default/inspec/repo_spec.rb
+++ b/test/integration/default/inspec/repo_spec.rb
@@ -1,4 +1,9 @@
-describe 'Debian Apt Repo' do
+if os.family == 'redhat'
+  describe yumrepo('cassandra') do
+    it { should exist }
+    it { should be_enabled }
+  end
+else
   describe file('/etc/apt/sources.list.d/cassandra.sources.list') do
     it { should exist }
     it { should be_file }