--- a/test/integration/default/inspec/repo_spec.rb
+++ b/test/integration/default/inspec/repo_spec.rb
@@ -1,11 +1,11 @@
-if os.family == 'redhat'
-  describe yum.repo('cassandra') do
-    it { should exist }
-    it { should be_enabled }
-  end
-else
+if os.family == 'debian'
   describe file('/etc/apt/sources.list.d/cassandra.sources.list') do
     it { should exist }
     it { should be_file }
+  end
+else
+  describe yum.repo('cassandra') do
+    it { should exist }
+    it { should be_enabled }
   end
 end