--- a/test/integration/default/inspec/service_spec.rb
+++ b/test/integration/default/inspec/service_spec.rb
@@ -1,5 +1,3 @@
-require_relative '../spec_helper'
-
 describe 'Common Service' do
   describe service('cassandra') do
     it { should be_enabled }