--- a/test/integration/default/inspec/service_spec.rb
+++ b/test/integration/default/inspec/service_spec.rb
@@ -1,6 +1,5 @@
-describe 'Common Service' do
-  describe service('cassandra') do
-    it { should be_enabled }
-    it { should be_running }
-  end
+describe service('cassandra') do
+  it { should be_enabled }
+  it { should be_running }
+  it { should be_installed }
 end