--- a/test/integration/default/inspec/service_spec.rb
+++ b/test/integration/default/inspec/service_spec.rb
@@ -1,5 +1,13 @@
-describe service('cassandra') do
-  it { should be_enabled }
-  it { should be_running }
-  it { should be_installed }
+if os.family == 'redhat'
+  describe service('cassandra') do
+    it { should be_enabled }
+    it { should be_running }
+    it { should be_installed }
+  end
+else
+  describe upstart_service('cassandra') do
+    it { should be_enabled }
+    it { should be_running }
+    it { should be_installed }
+  end
 end