--- a/test/spec/debian/config_file_spec.rb
+++ b/test/spec/debian/config_file_spec.rb
@@ -1,5 +1,3 @@
-require_relative '../spec_helper'
-
 describe 'Debian Configuration File' do
   describe file('/etc/cassandra/cassandra.yaml') do
     it { should be_file }