--- a/test/spec/debian/config_file_spec.rb
+++ b/test/spec/debian/config_file_spec.rb
@@ -1,6 +1,0 @@
-describe 'Debian Configuration File' do
-  describe file('/etc/cassandra/cassandra.yaml') do
-    it { should be_file }
-    it { should contain 'cluster_name: MyCassandraCluster' }
-  end
-end