--- a/test/spec/redhat/config_file_spec.rb
+++ b/test/spec/redhat/config_file_spec.rb
@@ -1,5 +1,3 @@
-require_relative '../spec_helper'
-
 describe 'RedHat Configuration File' do
   describe file('/etc/cassandra/conf/cassandra.yaml') do
     it { should be_file }