--- a/test/spec/redhat/config_file_spec.rb
+++ b/test/spec/redhat/config_file_spec.rb
@@ -1,6 +1,0 @@
-describe 'RedHat Configuration File' do
-  describe file('/etc/cassandra/conf/cassandra.yaml') do
-    it { should be_file }
-    it { should contain 'cluster_name: MyCassandraCluster' }
-  end
-end