--- a/test/spec/redhat/rackdc_spec.rb
+++ b/test/spec/redhat/rackdc_spec.rb
@@ -1,5 +1,3 @@
-require_relative '../spec_helper'
-
 describe 'RedHat Rack/DC Configuration File' do
   describe file('/etc/cassandra/default.conf/cassandra-rackdc.properties') do
     it { should be_file }