--- a/test/spec/redhat/repo_spec.rb
+++ b/test/spec/redhat/repo_spec.rb
@@ -1,5 +1,3 @@
-require_relative '../spec_helper'
-
 describe 'Red Hat Yum Repo' do
   describe yumrepo('cassandra') do
     it { should exist }