--- a/test/spec/redhat/repo_spec.rb
+++ b/test/spec/redhat/repo_spec.rb
@@ -1,6 +1,0 @@
-describe 'Red Hat Yum Repo' do
-  describe yumrepo('cassandra') do
-    it { should exist }
-    it { should be_enabled }
-  end
-end