--- a/test/spec/spec_helper.rb
+++ b/test/spec/spec_helper.rb
@@ -1,2 +1,0 @@
-require 'serverspec'
-set :backend, :exec