… test/ansible-review/standards.py
… test/integration/default/inspec/node_status_spec.rb
… test/inventory/group_vars/cassandra.yml
… test/inventory/hosts
… test/spec/common/cluster_name_spec.rb
… test/spec/common/directories_spec.rb
… test/spec/common/package_spec.rb
… test/spec/common/service_spec.rb
… test/test.yml