The previous non-optimised build took 12m 10s to run (total running time of 48m 58s) [1].  This last trial build took 6m 4s (total running time of 26m 40s) [2].  This was achieved by:

* Using packages instead of pip modules where possible
* Migrating from Serverspec to InSpec [3]
* Combining TravisCI jobs to reduce repetition

[1] - https://travis-ci.org/locp/ansible-role-cassandra/builds/467527321
[2] - https://travis-ci.org/locp/ansible-role-cassandra/builds/467770627
[3] - https://www.inspec.io/docs/reference/migration/